backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.3
scipy==1.11.4
imbalanced-learn==0.11.0
anthropic==0.20.0
python-dotenv==1.0.0
pydantic==2.5.0
joblib==1.3.2
